+++ product-service/src/main/java/com/ecommerce/product_service/entities/Product.java
@@ -4,6 +4,8 @@ package com.ecommerce.product_service.entities;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
+
 @Document(value = "products")
 public class Product {
 
@@ -11,13 +13,13 @@ public class Product {
     private String id;
     private String name;
     private String description;
-    private Double price;
+    private BigDecimal price;
     private String category;
 
     public Product() {
     }
 
-    public Product(String id, String name, String description, Double price, String category) {
+    public Product(String id, String name, String description, BigDecimal price, String category) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -51,11 +53,11 @@ public class Product {
         this.description = description;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
